requirements.txt
asgiref==3.2.10
beautifulsoup4==4.9.3
certifi==2020.6.20
cloudinary==1.22.0
dj-database-url==0.5.0
Django==3.1.2
django-bootstrap4==2.3.1
django-heroku==0.3.1
django-tinymce==3.1.0
djangorestframework==3.12.1
djangorestframework-simplejwt==4.4.0
importlib-metadata==2.0.0
Pillow==8.0.1
psycopg2==2.8.6
PyJWT==1.7.1
python-decouple==3.3
pytz==2020.1
six==1.15.0
soupsieve==2.0.1
sqlparse==0.4.1
urllib3==1.25.11
whitenoise==5.2.0
zipp==3.4.0
